+++ Changes.md
@@ -49,6 +49,7 @@ Version 2.9.1.6
  45. Remove redundant includes in auto-loaded classes.
  46. Add 'Style Guide' admin setting page.
      Can also be accessed via the URL '//yourmoodlesite/theme/shoehorn/pages/styleguide.php' without logging on.
+ 47. Add 'Format responsive' overrides for the 'Columns format'.
 
 Version 2.9.1.5
   1. MDL-51921.